Initial commit
Started working on the organization of the application.  Added basic
controllers for the login page and channel, which seem to be functioning.  A
user service to maintain the status of the currently logged in user is in the
works.

Also added the old favicon and CSS as those can stay for a while until we get
SASS up and running.