Fix a few scrollbar issues
This fixes the scrollbar having a white background when displayed
on the document.  The background color of the scrollbar is now
inherited from the background color of the bounding box.  To ensure
that this will always grab the right color, `background-color` was
also added to the <html> element.

This also removes the scrollbar buttons completely, as they were
previously hidden.  The scrollbar thumb now extends to the edges of
the container, so it doesn't look as strange when it stops short.
While this reduces the usability of the scrollbar, it is better
than what was previously implemented.

The invalid background colors (`none`) were changed to `transparent`
as well.  `none` is not a valid background color, and webkit was
falling back to `transparent` anyway.