Only show nick changes in channels with user present
This fixes #66 by checking if the user who the NICK message is
for is located within each channel.  This fixes an issue where the
message would be broadcast to all channels that the current user
was present in, even if the user that changed their name was not
actually present.